backend/requirements_minimal.txt
fastapi==0.104.1
mangum==0.17.0
boto3==1.34.0
PyJWT==2.8.0
bcrypt==4.1.2
pydantic==2.5.0
aiohttp==3.12.15
google-auth==2.28.1
requests==2.31.0